+++ packages/ckeditor5-admonition/src/admonitioncommand.ts
